Dcwl 2507 customs-notification] Scala 3 Upgrade (#212)
* DCWL-2507 Scala 3 upgrade Initial commit

* DCWL-2507 Scala 3 upgrade fixing test case failure

* DCWL-2507 Scala 3 upgrade fixing test case failure

* DCWL-2507 Scala 3 upgrade fixing test case failure

* DCWL-2507 Scala 3 upgrade fixing test case failure

* DCWL-2507 Scala 3 upgrade resolving warnings

* DCWL-2507 Scala 3-Increased Statement coverage

* DCWL-2507 Scala 3-Increased Statement coverage

* DCWL-2507 Scala 3-Addressed Warnings

* DCWL-2507 Scala 3-org.playframework:sbt-plugin to 3.0.8

* DCWL-2507 Scala 3-org.playframework:sbt-plugin to 3.0.8

---------

+++ app/uk/gov/hmrc/customs/notification/repo/NotificationWorkItemRepo.scala
@@ -32,7 +32,6 @@ import uk.gov.hmrc.mongo.play.json.Codecs
 import uk.gov.hmrc.mongo.workitem.ProcessingStatus.{Failed, InProgress, PermanentlyFailed}
 import uk.gov.hmrc.mongo.workitem.{ProcessingStatus, ResultStatus, WorkItem, WorkItemRepository}
 import uk.gov.hmrc.mongo.{MongoComponent, MongoUtils}
-
 import java.time.{Duration, Instant, ZonedDateTime}
 import java.util.UUID
 import java.util.concurrent.TimeUnit
@@ -101,7 +100,7 @@ class NotificationWorkItemMongoRepo @Inject()(mongo: MongoComponent,
           indexOptions = IndexOptions()
             .name(TTL_INDEX_NAME)
             .unique(false)
-            .expireAfter(ttlInSeconds, TimeUnit.SECONDS)
+            .expireAfter(ttlInSeconds.toLong, TimeUnit.SECONDS)
         ),
         IndexModel(
           keys = descending("clientNotification.clientId"),
